fix(warranty): upload plate image using cloud function instead of cloud file id
Plate number image was uploaded to cloud storage and file ID was then passed to cloud function.
Cloud function download the image with the given file ID and then send to Baidu AI. There seems to
be some synchronization problem where the cloud function will download the original file which is
meant to be replaced by the new file. The routine now change to upload the file in the cloud
function and send to Baidu AI simultaneously, and then return the file ID.